closes #14 - add modal to confirm deletion of seller

+++ sellers-frontend/src/app/components/sellers-table/sellers-table.component.ts
@@ -1,6 +1,7 @@
 import { Component, EventEmitter, input, Input, Output } from '@angular/core';
 import { SellerService } from '../../services/seller.service';
 import Seller from '../../interfaces/Seller';
+import { NgbModal } from '@ng-bootstrap/ng-bootstrap';
 
 @Component({
   selector: 'app-sellers-table',
@@ -8,9 +9,12 @@ import Seller from '../../interfaces/Seller';
   styleUrl: './sellers-table.component.css'
 })
 export class SellersTableComponent {
-  constructor(private sellerService: SellerService) { }
+  constructor(private sellerService: SellerService,
+    private modalService: NgbModal
+  ) { }
 
   seller: Seller = {} as Seller;
+  sellerToDelete: Seller = {} as Seller;
   sellers: Seller[] = []
   register = false;
 
@@ -58,9 +62,14 @@ export class SellersTableComponent {
     this.register = true;
   }
 
-  delete(seller: Seller) {
-    this.sellerService.deleteSeller(seller).subscribe({
-      next: () => this.getSellers()
+  delete(modal: any, seller: Seller) {
+    this.sellerToDelete = seller;
+    this.modalService.open(modal).result.then((confirm) => {
+      if (confirm) {
+        this.sellerService.deleteSeller(seller).subscribe({
+          next: () => this.getSellers()
+        })
+      }
     })
   }
